requirements.txt
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
google-cloud-storage>=2.10.0
feedgen>=0.9.0
PyYAML>=6.0
requests>=2.31.0
